requirements.txt
Flask==2.2.3
Flask_Bootstrap==3.3.7.1
Flask_WTF==1.1.1
Pillow==10.0.0
WTForms==3.0.1
